feat(contract): rename interface views to contracts
