+++ README.md
@@ -1,45 +1,66 @@
-[![Build Status](https://travis-ci.org/moseskamira/AndroidLevelupProject.svg?branch=develop)](https://travis-ci.org/moseskamira/AndroidLevelupProject)
-[![Maintainability](https://api.codeclimate.com/v1/badges/79307512581ddef1b841/maintainability)](https://codeclimate.com/github/moseskamira/AndroidLevelupProject/maintainability)
-[![codecov](https://codecov.io/gh/moseskamira/AndroidLevelupProject/branch/develop/graph/badge.svg)](https://codecov.io/gh/moseskamira/AndroidLevelupProject)
+# 🚀 Android LevelUp Project
 
-# AndroidLevelupProject
-The project involves building an Android App to retrieve a list of Java Developers in Nairobi using Github API
-Also one should be able to view details of a developer's profile as well as sharing the developer's profile on their favorite sites or social media.
+<p align="center">
+  <img src="https://img.shields.io/badge/Android-Kotlin-3DDC84?style=for-the-badge&logo=android" />
+  <img src="https://img.shields.io/badge/Architecture-MVP-blue?style=for-the-badge" />
+  <img src="https://img.shields.io/badge/API-GitHub-black?style=for-the-badge&logo=github" />
+  <img src="https://img.shields.io/badge/Platform-Android-green?style=for-the-badge" />
+</p>
 
+<p align="center">
+  <b>A clean Android app that displays GitHub developers in Nairobi using the GitHub API.</b>
+</p>
 
-# Getting Started
-Follow the Instructions below to have the Application running on your machine..
+---
 
-## Prerequisites
-Android Studio <br/>
-JDK <br/>
-Git <br/>
-### Installation
-Clone the repo on the link: https://github.com/moseskamira/AndroidLevelupProject.git <br/>
-Import the project into android studio <br/>
-Create an android virtual device (AVD) and have it started <br/>
-Run the project <br/>
-Run the tests <br/>
-In the terminal, run ./gradlew jacocoTestCoverage to run tests and generate a coverage report
+## 📱 Overview
 
+Android LevelUp Project is a native Android application that demonstrates real-world Android development using the **GitHub API**.
 
-## MOCK UPS
-Below is a brief description of the mockups designed for the Android Application and tool used.
-### Selected Tool for designing Mockups
-JustInMind
+It allows users to:
+- 👨‍💻 Browse developers in Nairobi
+- 🔍 View detailed developer profiles
+- 📤 Share developer profiles via Android intents
 
-### Reasons for selecting the Tool
-The tool can be used locally without necessarily having internet connectivity <br/>
-Avails navigable interfaces that require no much user experience
+Built with a focus on **MVP architecture**, clean code, and scalability.
 
+---
 
+## ✨ Features
 
-### Link To Navigable Mock ups
-https://www.justinmind.com/usernote/tests/39777996/39783299/39783301/index.html#/screens/d12245cc-1680-458d-89dd-4f0d7fb22724
+- 📡 Fetch developers from GitHub API
+- 📃 Display users in Linear & Grid RecyclerView
+- 👤 View detailed profile screen
+- 📤 Share developer profile
+- 🎨 Clean UI design
+- 🧱 MVP architecture
 
-#### Guideline to using the navigable link:
-Click on the github name of the first sample developer in the names of 'moseskamira' <br/>
-After accessing his profile details, you can click 'SHARE' to share
+---
+
+## 🧠 Architecture (MVP)
+
+### Layers
+
+- **Model**
+    - DTO (API response models)
+    - Entity (Domain models)
+    - Repository (Data handling)
+    - Mapper (DTO ↔ Entity)
+
+- **View**
+    - Activities
+    - Fragments
+    - Adapters
+
+- **Presenter**
+    - Business logic
+    - API coordination
+    - UI state updates
+
+---
+
+### 🔁 Data Flow
+
 
 
 ### Linear Display of List of All Developers In Nairobi
