+++ app/src/androidTest/java/com/example/androidcodelabapp/view/MainActivityTest.java
@@ -2,6 +2,7 @@ package com.example.androidcodelabapp.view;
 
 
 import com.example.androidcodelabapp.R;
+import com.example.androidcodelabapp.view.activities.UsersActivity;
 
 import org.junit.After;
 import org.junit.Rule;
