+++ app/src/main/java/com/example/androidcodelabapp/model/data/APIService.kt
@@ -1,10 +1,11 @@
-package com.example.androidcodelabapp.service
+package com.example.androidcodelabapp.model.data
 
+import com.example.androidcodelabapp.util.Constants
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
 object APIService {
-    private val retrofit: Retrofit = Retrofit.Builder().baseUrl("https://api.github.com/")
+    private val retrofit: Retrofit = Retrofit.Builder().baseUrl(Constants.BASEURL)
         .addConverterFactory(GsonConverterFactory.create()).build()
     val apiClient: APIClient = retrofit.create(APIClient::class.java)
