+++ app/src/main/java/com/example/androidcodelabapp/model/domain/entities/GithubUsersResponse.kt
@@ -1,0 +1,8 @@
+package com.example.androidcodelabapp.model.domain.entities
+
+import com.google.gson.annotations.SerializedName
+
+data class GithubUsersResponse(
+    @SerializedName("items")
+    val githubUsers: ArrayList<GithubUser> = arrayListOf()
+)
