--- app/src/main/java/com/example/androidcodelabapp/view/AllDevelopersContract.kt
@@ -1,7 +1,0 @@
-package com.example.androidcodelabapp.view;
-
-import com.example.androidcodelabapp.model.GithubUsersResponse;
-
-interface AllDevelopersContract {
-    fun showDevelopers(response: GithubUsersResponse)
-}
