--- app/src/main/java/com/example/androidcodelabapp/view/SingleDeveloperContract.kt
@@ -1,7 +1,0 @@
-package com.example.androidcodelabapp.view
-
-import com.example.androidcodelabapp.model.GithubUser
-
-interface SingleDeveloperContract {
-    fun showDeveloperProfile(profile: GithubUser)
-}
