+++ app/src/main/java/com/example/androidcodelabapp/view/contracts/AllDevelopersContract.kt
@@ -1,0 +1,7 @@
+package com.example.androidcodelabapp.view.contracts;
+
+import com.example.androidcodelabapp.model.domain.entities.GithubUsersResponse;
+
+interface AllDevelopersContract {
+    fun showDevelopers(response: GithubUsersResponse)
+}
