+++ app/src/main/java/com/example/androidcodelabapp/view/contracts/SingleDeveloperContract.kt
@@ -1,0 +1,7 @@
+package com.example.androidcodelabapp.view.contracts
+
+import com.example.androidcodelabapp.model.domain.entities.GithubUser
+
+interface SingleDeveloperContract {
+    fun showDeveloperProfile(profile: GithubUser)
+}
