+++ app/src/main/res/layout/activity_users.xml
@@ -5,7 +5,7 @@
     android:id="@+id/cordinator"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.UsersActivity">
+    tools:context=".view.activities.UsersActivity">
 
     <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
         android:id="@+id/swipe"
